// frontend-ksiegarnia/src/app/Model/dane.model.ts
export class Dane {
  userdataId: number;
  name: string | null;
  surname: string | null;
  mobilePhone: string | null;
  city: string | null;
  postcode: string | null;
  province: string | null;
  houseApartmentNr: string | null;
  street: string | null;
  cardNumber: string | null;
  expirationDate: string | null;
  cvvnumber: string | null;
  type: string;
  userData: number;
}